misc/HtmlTestSuite/results/test-0028.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>
'use strict';
setup({ single_test: true });

function iframeOnLoad() {
  const target = document.querySelector('#target');
  const wt = target.contentWindow.wt;
  target.remove();
  const streams = wt.incomingBidirectionalStreams;
  assert_equals(typeof streams, 'object', 'streams should be an object');
  done();
}
</script>

</head><body><iframe id="target" onload="iframeOnLoad()" srcdoc="&#xA;&lt;!doctype html&gt;&#xA;&lt;script src=/common/get-host-info.sub.js&gt;&lt;/script&gt;&#xA;&lt;script src=resources/webtransport-test-helpers.sub.js&gt;&lt;/script&gt;&#xA;&lt;script&gt;&#xA;window.wt = new WebTransport(webtransport_url('echo.py'));&#xA;&lt;/script&gt;&#xA;"/>
</body></html>
